move deps to separate file and add dialer
