+++ modules/renter/hostdb/hostdb.go
@@ -27,45 +27,14 @@ var (
 	errNilTpool  = errors.New("cannot create hostdb with nil transaction pool")
 )
 
-// These interfaces define the HostDB's dependencies. Using the smallest
-// interface possible makes it easier to mock these dependencies in testing.
-type (
-	hdbConsensusSet interface {
-		ConsensusSetSubscribe(modules.ConsensusSetSubscriber)
-	}
-	hdbTransactionBuilder interface {
-		AddArbitraryData([]byte) uint64
-		AddFileContract(types.FileContract) uint64
-		Drop()
-		FundSiacoins(types.Currency) error
-		Sign(bool) ([]types.Transaction, error)
-		View() (types.Transaction, []types.Transaction)
-	}
-	hdbWallet interface {
-		NextAddress() (types.UnlockConditions, error)
-		StartTransaction() hdbTransactionBuilder
-	}
-	hdbTransactionPool interface {
-		AcceptTransactionSet([]types.Transaction) error
-	}
-)
-
-// because hdbWallet is not directly compatible with modules.Wallet (differing
-// type signatures for StartTransaction), we must provide a shim type.
-type hdbWalletShim struct {
-	w modules.Wallet
-}
-
-func (ws *hdbWalletShim) NextAddress() (types.UnlockConditions, error) { return ws.w.NextAddress() }
-func (ws *hdbWalletShim) StartTransaction() hdbTransactionBuilder      { return ws.w.StartTransaction() }
-
 // The HostDB is a database of potential hosts. It assigns a weight to each
 // host based on their hosting parameters, and then can select hosts at random
 // for uploading files.
 type HostDB struct {
-	// modules
+	// dependencies
 	wallet hdbWallet
 	tpool  hdbTransactionPool
+	dialer hdbDialer
 
 	// The hostTree is the root node of the tree that organizes hosts by
 	// weight. The tree is necessary for selecting weighted hosts at
@@ -119,7 +88,7 @@ func New(cs hdbConsensusSet, wallet modules.Wallet, tpool hdbTransactionPool, pe
 		return nil, errNilTpool
 	}
 
-	hdb, err := newHostDB(&hdbWalletShim{w: wallet}, tpool, persistDir)
+	hdb, err := newHostDB(&hdbWalletShim{w: wallet}, tpool, stdDialer{}, persistDir)
 	if err != nil {
 		return nil, err
 	}
@@ -142,10 +111,11 @@ func New(cs hdbConsensusSet, wallet modules.Wallet, tpool hdbTransactionPool, pe
 // newHostDB creates a HostDB using the provided dependencies. It does not
 // have any side effects (i.e. it does not spawn background threads, perform
 // I/O, or call stateful methods of its dependencies.)
-func newHostDB(wallet hdbWallet, tpool hdbTransactionPool, persistDir string) (*HostDB, error) {
+func newHostDB(w hdbWallet, tpool hdbTransactionPool, d hdbDialer, persistDir string) (*HostDB, error) {
 	hdb := &HostDB{
-		wallet: wallet,
+		wallet: w,
 		tpool:  tpool,
+		dialer: d,
 
 		contracts:   make(map[types.FileContractID]hostContract),
 		activeHosts: make(map[modules.NetAddress]*hostNode),
